validate hex binary
