+++ proof-parser/src/layout.rs
@@ -3,8 +3,6 @@ use std::{collections::BTreeMap, convert::TryInto, fmt::Display};
 use num_bigint::BigUint;
 use serde::Deserialize;
 
-use crate::proof_params::{Fri, ProofParameters};
-
 // For now only the recursive and starknet layouts is supported
 #[derive(Debug, Clone, Copy, PartialEq, Deserialize)]
 #[serde(rename_all = "snake_case")]
@@ -156,124 +154,4 @@ impl Layout {
             _ => unimplemented!(),
         }
     }
-}
-
-// https://github.com/cartridge-gg/stone-prover/blob/fd78b4db8d6a037aa467b7558ac8930c10e48dc1/src/starkware/stark/stark.cc#L303-L304
-#[cfg(test)]
-pub fn fri_degree_bound(proof_params: &ProofParameters) -> u32 {
-    let mut expected = proof_params.stark.fri.last_layer_degree_bound;
-    for s in &proof_params.stark.fri.fri_step_list {
-        expected *= 1 << s
-    }
-    expected
-}
-
-pub fn leaves(proof_params: &ProofParameters) -> Vec<usize> {
-    proof_params
-        .stark
-        .fri
-        .fri_step_list
-        .iter()
-        .skip(1)
-        .map(|&x| (1u32 << (x + 4)) - 16)
-        .map(|x| x as usize)
-        .collect()
-}
-
-// https://github.com/cartridge-gg/stone-prover/blob/fd78b4db8d6a037aa467b7558ac8930c10e48dc1/src/starkware/commitment_scheme/packaging_commitment_scheme.cc#L245-L250
-pub fn authentications() -> usize {
-    256 + 1
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct ProofStructure {
-    pub first_layer_queries: usize,
-    pub layer_count: usize,
-    pub composition_decommitment: usize,
-    pub oods: usize,
-    pub composition_leaves: usize,
-    pub last_layer_degree_bound: usize,
-    pub authentications: usize,
-    pub layer: Vec<usize>,
-    pub witness: Vec<usize>,
-}
-
-impl ProofStructure {
-    pub fn new(proof_params: &ProofParameters, layout: Layout, _n_steps: u32) -> Self {
-        let n_queries = proof_params.stark.fri.n_queries;
-        let mask_len = layout.mask_len();
-        let layout = layout.get_consts();
-
-        ProofStructure {
-            // https://github.com/cartridge-gg/stone-prover/blob/fd78b4db8d6a037aa467b7558ac8930c10e48dc1/src/starkware/stark/stark.cc#L276-L277
-            first_layer_queries: (n_queries * layout.num_columns_first) as usize,
-
-            layer_count: proof_params.stark.fri.fri_step_list.len() - 1,
-            composition_decommitment: (n_queries * layout.num_columns_second) as usize,
-
-            // https://github.com/cartridge-gg/stone-prover/blob/fd78b4db8d6a037aa467b7558ac8930c10e48dc1/src/starkware/stark/oods.cc#L92-L93
-            oods: mask_len + layout.num_columns_second as usize - 1,
-            last_layer_degree_bound: proof_params.stark.fri.last_layer_degree_bound as usize,
-
-            // https://github.com/cartridge-gg/stone-prover/blob/fd78b4db8d6a037aa467b7558ac8930c10e48dc1/src/starkware/stark/composition_oracle.cc#L288-L289
-            composition_leaves: 2 * n_queries as usize,
-            authentications: authentications(),
-
-            layer: leaves(proof_params),
-            witness: witness(&proof_params.stark.fri),
-        }
-    }
-}
-
-pub fn witness(fri: &Fri) -> Vec<usize> {
-    let first_fri_step = 16;
-    let mut cumulative = 0;
-    let mut vec = Vec::new();
-
-    // https://github.com/cartridge-gg/stone-prover/blob/fd78b4db8d6a037aa467b7558ac8930c10e48dc1/src/starkware/fri/fri_details.cc#L93-L97
-    for v in fri.fri_step_list.iter().skip(1) {
-        cumulative += *v;
-        vec.push(first_fri_step - cumulative);
-    }
-
-    // https://github.com/cartridge-gg/stone-prover/blob/fd78b4db8d6a037aa467b7558ac8930c10e48dc1/src/starkware/fri/fri_details.cc#L74-L82
-    vec.into_iter()
-        .map(|len| fri.n_queries * len + 1)
-        .map(|x| x as usize)
-        .collect()
-}
-
-#[test]
-fn test_lens() {
-    let n_steps = 16384;
-    let layout = Layout::Recursive;
-    let proof_params = ProofParameters {
-        stark: crate::proof_params::Stark {
-            fri: Fri {
-                fri_step_list: vec![0, 4, 4, 3],
-                last_layer_degree_bound: 128,
-                n_queries: 16,
-                proof_of_work_bits: 30,
-            },
-            log_n_cosets: 3,
-        },
-        n_verifier_friendly_commitment_layers: 0,
-    };
-
-    assert_eq!(fri_degree_bound(&proof_params), 262144);
-
-    let result = ProofStructure::new(&proof_params, layout, n_steps);
-
-    let expected = ProofStructure {
-        first_layer_queries: 112,
-        layer_count: 3,
-        composition_decommitment: 48,
-        oods: 135,
-        last_layer_degree_bound: 128,
-        composition_leaves: 32,
-        authentications: 257,
-        layer: vec![240, 240, 112],
-        witness: vec![193, 129, 81],
-    };
-    assert_eq!(result, expected);
 }
